move ssh funcs into a new pkg
